Now store is working with 2 reducers, cart and books
