+++ react/finger-exercises/redux/02 - react-redux/src/redux/store.js
@@ -1,6 +1,7 @@
-import { createStore, compose } from 'redux';
+import { createStore, compose, combineReducers } from 'redux';
 
 import books from './book/reducer';
+import cart from './cart/reducer';
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose; // eslint-disable-line no-underscore-dangle
-export default createStore(books, composeEnhancers());
+export default createStore(combineReducers(books, cart), composeEnhancers());
